Move MouseCommandFactory into a separated file
